+++ src/parser/data-structure/object.ts
@@ -1,6 +1,5 @@
 import { createDataStructureParser } from '../parser';
-import { getPossibleSelfDefinedType, isEqual } from '../util';
-import { freeze } from './util';
+import { getPossibleSelfDefinedType } from '../util';
 
 type ObjectParams<T> = Readonly<{
     u: unknown;
@@ -9,31 +8,22 @@ type ObjectParams<T> = Readonly<{
 
 const parseAsAnObject = <T extends Object>(
     u: ObjectParams<T>['u'],
-    t: ObjectParams<T>['t'],
-    isImmutable: boolean
-) => {
-    const actualType = getPossibleSelfDefinedType(u);
-    return {
-        ...createDataStructureParser<T>({
-            actualType,
-            expectedType: 'object',
-            t: () => t(u),
-        }),
-        orElseGetEmptyObject: () =>
-            !isEqual('object', actualType)
-                ? {}
-                : freeze(() => t(u), isImmutable),
-    };
-};
+    t: ObjectParams<T>['t']
+) =>
+    createDataStructureParser<T>({
+        expectedType: 'object',
+        actualType: getPossibleSelfDefinedType(u),
+        t: () => t(u),
+    });
 
 const parseAsObject = <T extends Object>(
     u: ObjectParams<T>['u'],
     t: ObjectParams<T>['t']
-) => parseAsAnObject(u, t, false);
+) => parseAsAnObject(u, t);
 
 const parseAsReadonlyObject = <T extends Object>(
     u: ObjectParams<T>['u'],
     t: ObjectParams<T>['t']
-) => parseAsAnObject<Readonly<T>>(u, t, true);
+) => parseAsAnObject<Readonly<T>>(u, t);
 
 export { parseAsObject, parseAsReadonlyObject };