+++ test/primitive/string.ts
@@ -10,28 +10,19 @@ const stringParserTest = () => {
                 value: true,
             } as const,
         ])('data => %p', ({ value }) => {
-            const parseString = parseAsString(value);
-            expect(parseString.orElseGetEmptyString()).toEqual('');
-
-            const exact = parseString.exactlyAs('12');
-            expect(exact.orElseGetEmptyString()).toEqual('');
+            expect(parseAsString(value).elseGet('')).toEqual('');
+            expect(parseAsString(value).elseLazyGet(() => undefined)).toEqual(
+                undefined
+            );
         });
     });
 
-    describe('Test parse as string get empty string', () => {
-        test.each([
-            {
-                value: '123',
-            } as const,
-            {
-                value: 'Yep',
-            } as const,
-        ])('data => %p', ({ value }) => {
-            const parseString = parseAsString(value);
-            expect(parseString.orElseGetEmptyString()).toEqual(value);
-
-            const exact = parseString.exactlyAs(value);
-            expect(exact.orElseGetEmptyString()).toEqual(value);
+    describe('Test parse as string get', () => {
+        test.each(['', '1', '2'])('data => %p', (value) => {
+            expect(parseAsString(value).elseGet('')).toEqual(value);
+            expect(parseAsString(value).elseLazyGet(() => undefined)).toEqual(
+                value
+            );
         });
     });
 };